refactor: pattern sugestion
