+++ src/routes/user.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 
 import * as UserController from '../controllers/user';
-import { createUserSchema } from '../utils/user/validators';
+import { createUserSchema } from '../utils/validators/user/user.validator';
 import validatorMiddleware from '../utils/middleware/validator';
 
 const router = Router();
